hiding which stderr

+++ bin/install-datafaqs-dependencies.sh
@@ -93,24 +93,28 @@ function offer_install_aptget { # NOTE: This adds "dryrun" so that Prizms can in
    for package in $packages; do
       echo "The package $package is required to"
       echo "$reason."
-      already_there=`dpkg -l | grep $package` # See what is available: apt-cache search libapache2-mod
-      if [[ -z "$already_there" ]]; then 
-         echo "The $package package needs to be installed, which can be done with the following command:"
+      if [[ `which dpkg 2> /dev/null` && `which apt-get 2> /dev/null` ]]; then
+         already_there=`dpkg -l | grep $package` # See what is available: apt-cache search libapache2-mod
+         if [[ -z "$already_there" ]]; then 
+            echo "The $package package needs to be installed, which can be done with the following command:"
+            echo 
+            echo "$TODO $sudo apt-get install $package"
+            echo 
+            if [ "$dryrun" != "true" ]; then
+               read -p "Q: May we install the package $package using the command above? [y/n] " -u 1 install_it
+               if [[ "$install_it" == [yY] ]]; then 
+                  echo $sudo apt-get install $package
+                       $sudo apt-get install $package
+                  installed=1
+               fi   
+            fi
+         else 
+            echo "($package is already installed)"
+         fi   
          echo 
-         echo "$TODO $sudo apt-get install $package"
-         echo 
-         if [ "$dryrun" != "true" ]; then
-            read -p "Q: May we install the package $package using the command above? [y/n] " -u 1 install_it
-            if [[ "$install_it" == [yY] ]]; then 
-               echo $sudo apt-get install $package
-                    $sudo apt-get install $package
-               installed=1
-            fi   
-         fi
-      else 
-         echo "($package is already installed)"
-      fi   
-      echo 
+      else
+         echo "WARNING: handle non-apt-get"
+      fi
    done 
    return $installed
 }
